requirements.txt
confusable-homoglyphs==3.0.0
dj-database-url==0.4.2
Django==1.11
django-bootstrap3==9.1.0
django-registration==2.4.1
django-tinymce==2.7.0
gunicorn==19.7.1
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-decouple==3.1
pytz==2018.3
whitenoise==3.3.1
